Squashed commit of the following:

commit 47a52cbd8262a251ad85661da3a8b58dbb2addfa
Author: Sergey Fionov <fionov@gmail.com>
Date:   Fri Mar 27 16:32:48 2020 +0300

    Fix varargs passing on Android

+++ platform/android/dnsproxy/lib/src/main/cpp/android_dnsproxy.cpp
@@ -337,7 +337,7 @@ ag::dnsfilter::engine_params ag::android_dnsproxy::marshal_filter_params(JNIEnv 
     ag::dnsfilter::engine_params params;
 
     jint size = env->CallIntMethod(java_filter_params, size_method);
-    for (int64_t i = 0; i < size; ++i) {
+    for (jint i = 0; i < size; ++i) {
         jint key = env->CallIntMethod(java_filter_params, key_at_method, i);
         local_ref val(env, env->CallObjectMethod(java_filter_params, val_at_method, i));
         m_utils.visit_string(env, val.get(), [&](const char *str, jsize len) {
